remove old candidates module
